+++ UnityProject/Assets/Materials/BlurryVisionMaterial.mat
@@ -58,7 +58,7 @@ Material:
         m_Offset: {x: 0, y: 0}
     m_Ints: []
     m_Floats:
-    - _BlurryStrength: 7
+    - _BlurryStrength: 14
     - _BumpScale: 1
     - _COLORBLIND_MODE: 1
     - _Cutoff: 0.5
