+++ UnityProject/Assets/Scripts/Systems/Electricity/FunctionsAndClasses/Current.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Systems.Electricity
+{
+	public class Current
+	{
+		public double current = 0;
+	}
+}
