+++ UnityProject/Assets/Scripts/Systems/Electricity/FunctionsAndClasses/Current.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4dcea60060ca480459243d208023dd29
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
