+++ UnityProject/Assets/Scripts/Systems/Electricity/FunctionsAndClasses/WrapCurrent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc1c665bbb93e3246945cb31d9aa6422
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
